socialmedia/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.40.0
cryptography==42.0.7
defusedxml==0.7.1
Deprecated==1.2.14
Django==5.0.6
django-allauth==0.61.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
django-sslify==0.2.7
djangorestframework==3.15.1
drf-yasg==1.21.7
environ==1.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
inflection==0.5.1
jwcrypto==1.5.6
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.2
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wrapt==1.16.0
